requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diagrams==0.20.0
Django==4.0.3
django-allauth==0.49.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-countries==7.2.1
django-discover-runner==1.0
django-extensions==3.1.5
django-filebrowser==4.0.1
django-grappelli==3.0.1
django-libsass==0.9
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
django-sass-compiler==1.1.0
django-tables2==2.4.1
django-webtest==1.9.10
docutils==0.18.1
entrypoints==0.3
et-xmlfile==1.1.0
execnet==1.9.0
factory-boy==3.2.1
Faker==13.3.2
fontawesomefree==6.0.0
gql==3.0.0
graphql-core==3.2.0
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.7.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
libsass==0.21.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
model-bakery==1.4.0
multidict==6.0.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.3
oauthlib==3.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phone-gen==2.2.3
phonenumbers==8.12.43
pickleshare==0.7.5
Pillow==9.0.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==2.0.10
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.11.0
pycares==4.1.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-cas==1.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
rcssmin==1.1.0
requests==2.27.1
requests-oauthlib==1.3.0
rjsmin==1.2.0
robotframework==5.0
robotframework-datadriver==1.6.1
robotframework-debuglibrary==2.2.2
robotframework-djangolibrary==3.1.0
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==6.0.0
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
username==2.1.0
verify-email==2.4.3
waitress==2.1.0
wcwidth==0.2.5
webdrivermanager==0.10.0
webencodings==0.5.1
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.0.3
whitenoise==5.3.0
wsproto==1.0.0
xlrd==2.0.1
yarl==1.7.2
